debug commit

+++ blaze/datadescriptor/tests/test_h5py_data_descriptor.py
@@ -6,15 +6,32 @@ import os
 from dynd import nd
 import h5py
 import numpy as np
-from sys import stdout
+import random
+
+from threading import Lock
+
+lock = Lock()
+
+def random_filename(ext):
+    fn = 'test-1.%s' % ext
+    while os.path.exists(fn):
+        fn = 'rand_%d.%s' % (random.randint(2, 100000), ext)
+    return fn
+
+
+def log(s):
+    with open('test.txt', 'a') as f:
+        f.write(s)
+        f.write('\n')
 
 
 class SingleTestClass(unittest.TestCase):
     def setUp(self):
-        self.filename = tempfile.mktemp('h5')
+        self.filename = random_filename('h5')
 
     def tearDown(self):
-        os.remove(self.filename)
+        if os.path.exists(self.filename):
+            os.remove(self.filename)
 
     """
     def test_creation(self):
@@ -28,7 +45,8 @@ class SingleTestClass(unittest.TestCase):
         """
 
     def test_existing_array(self):
-        stdout.flush()
+        log('test_existing_array')
+
         with h5py.File(self.filename, 'w') as f:
             d = f.create_dataset('data', (3, 3), dtype='i4',
                                  chunks=True, maxshape=(None, 3))
@@ -46,20 +64,8 @@ class SingleTestClass(unittest.TestCase):
 
         self.assertEquals(str(dd.dshape), 'var * 3 * int32')
 
-    def test_schema(self):
-        dd = H5PY_DDesc(self.filename, '/data', 'a', schema='2 * int32')
-
-        self.assertEquals(str(dd.schema), '2 * int32')
-        self.assertEquals(str(dd.dshape), 'var * 2 * int32')
-
-    def test_dshape(self):
-        dd = H5PY_DDesc(self.filename, '/data', 'a', dshape='var * 2 * int32')
-
-        self.assertEquals(str(dd.schema), '2 * int32')
-        self.assertEquals(str(dd.dshape), 'var * 2 * int32')
-
     def test_extend_chunks(self):
-        stdout.flush()
+        log('test_extend_chunks')
         with h5py.File(self.filename, 'w') as f:
             d = f.create_dataset('data', (3, 3), dtype='i4',
                                  chunks=True, maxshape=(None, 3))
@@ -79,18 +85,46 @@ class SingleTestClass(unittest.TestCase):
         self.assertEquals(nd.as_py(result), nd.as_py(expected))
 
     def test_iterchunks(self):
-        stdout.flush()
+        log('test_iterchunks')
         with h5py.File(self.filename, 'w') as f:
             d = f.create_dataset('data', (3, 3), dtype='i8')
             d[:] = 1
         dd = H5PY_DDesc(self.filename, '/data')
         assert all(isinstance(chunk, nd.array) for chunk in dd.iterchunks())
 
-    def test_extend(self):
-        dd = H5PY_DDesc(self.filename, '/data', 'a', schema='2 * int32')
+def test_extend():
+    log('test_extend')
+    with openfile('h5') as filename:
+        log(filename)
+        log('written to file')
+        dd = H5PY_DDesc(filename, '/data', 'a', schema='2 * int32')
+        log('test_extend')
         dd.extend([(1, 1), (2, 2)])
+        log('test_extend')
 
         results = list(dd)
 
-        self.assertEquals(nd.as_py(results[0]), [1, 1])
-        self.assertEquals(nd.as_py(results[1]), [2, 2])
+        assert nd.as_py(results[0]) == [1, 1]
+        assert nd.as_py(results[1]) == [2, 2]
+
+
+def test_schema():
+    with openfile('h5') as filename:
+        dd = H5PY_DDesc(filename, '/data', 'a', schema='2 * int32')
+
+        assert str(dd.schema) == '2 * int32'
+        assert str(dd.dshape) == 'var * 2 * int32'
+
+
+def test_dshape():
+    with openfile('h5') as filename:
+        log('test_dshape')
+        log(filename)
+
+        dd = H5PY_DDesc(filename, '/data2', 'w', dshape='var * 2 * int32')
+
+        log('after h5py')
+
+        assert str(dd.schema) == '2 * int32'
+        assert str(dd.dshape) == 'var * 2 * int32'
+
